GH-2377: move unit test to LuceneSailTest
Re-indexing apparently does not work correctly for Solr Index, thus
moving it to LuceneSailTest
